Refactoring out helper functions from integration tests
